requirements.txt
backcall==0.1.0
certifi==2019.9.11
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
dataclasses-json==0.5.2
decorator==4.4.0
defusedxml==0.7.0rc1
dj-database-url==0.5.0
Django==2.2.5
django-cors-headers==3.6.0
django-extensions==3.0.5
django-filter==2.3.0
django-heroku==0.3.1
django-oauth-toolkit==1.3.3
django-rename-app==0.1.2
djangorestframework==3.11.1
gunicorn==19.9.0
idna==2.8
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
marshmallow==3.10.0
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
oauthlib==3.1.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2==2.8.3
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.4.2
PyJWT==2.0.0
python3-openid==3.2.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.3.0
six==1.12.0
sqlparse==0.3.0
stringcase==1.2.0
traitlets==4.3.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.6
wcwidth==0.1.7
whitenoise==5.2.0
